requirements.txt
GitPython
Paste
WebOb
click
requests
webapp2
